chore: better dev experience with ts-node & mocha
Before we had tests importing src files from `../dist`, so in order for
that to work we needed to build first, and if we were on watch mode
we needed 2 builds to finally build test file.

Also, vscode went a little crazy and didn't show errors correctly.

Now, we make test files depend on `../src`. So the actual files.

Also, we don't require people to first build, then run tests. We use
`ts-node` to build "on the fly" while running tests.

A tale of two tsconfig.json: we have 2:
  - tsconfig.json: includes src & test folders, is what ts-node uses
  but also what vscode uses in it's TsService
  - tsconfig-build.json the actual build config we use to create dist/
  folder. In only shrinks the include. to use 'src/' (& this avoid
  having `dist/src` & `dist/test`)