fix: Filter
FilterOptions is recieved as a paramter of Filter, and then transformed.
But the types are wrong.

So a ParsedFilterOptions type was created, and a copy & transform was
made.

WIP: There are some broken tests that I yet don't know how to fix, since
the use a fakeHttpProvider mocking capabilities not yet explored by me.
help?